t1/requirements.txt
asgiref==3.10.0
astroid==3.3.11
dill==0.4.0
Django==4.2.25
django-cors-headers==4.9.0
djangorestframework==3.16.1
importlib_metadata==8.7.0
isort==6.1.0
mccabe==0.7.0
pillow==11.3.0
platformdirs==4.4.0
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
sqlparse==0.5.3
tomli==2.3.0
tomlkit==0.13.3
typing_extensions==4.15.0
zipp==3.23.0
